Update message translator contract
The message translator classes should return a dictionary with
at least one key named 'content'. The value of 'content' will be
delivered to Route.handler callable.

This structure will allow us to change the routing behavior of a message
based on some metadata in the future.

Update related classes and tests.