Add add_current_dir_to_syspath decorator and decorate import_callable utility
This is an enhancement because client code does not need to be installed
or set PYTHONPATH, otherwise a ImportError might be raised if a callable
are set to ROUTE_DEFAULT_HANDLER or similar (eg, any configuration that uses
import_callable).

We try to add the current directory assuming the client code will
start Loafer at the project root.
There's nothing we can do if they don't, we also should pay attention
to some side effects of this code.