Rename loafer.ext.aws.consumers to loafer.ext.aws.providers
The semantic of what is a consumer or provider is confuse.
With this change, objects that "provides" messages to be processed
will be called  a "Provider".

Previously, they were called consumers because they "consume" messages
from a source.

The loafer.ext.aws.Consumer is now loafer.ext.aws.SQSProvider.
A lot of changes but most of them are replacing s/consumer/provider/