Refactor BaseSNSClient: use wildcard to find the topic arn
boto3 does not offer a method to retrieve a single topic arn by name.
The previous solution were going to de-paginate all the topics available,
cache them and then, return the arn matching the topic name.

With using wildcards, we can "build" a valid arn for a topic-name.
It's possible to have the same topic in different regions, but it's
unlikely (or a bad practice anyway).